requirements.txt
Flask==2.2.2
google-api-python-client==2.56.0
google-api-core==2.8.2
google-auth==2.10.0
google-auth-httplib2==0.1.0
py-healthcheck==1.10.1
python-dotenv==0.20.0